Move decorator to another package
