+++ requirements.txt
@@ -22,3 +22,5 @@ requests==2.20.1
 web3==4.8.2
 websockets==6.0
 yara-python==3.7.0
+base58==0.2.5
+pycrypto==2.6.1
